examples: Reduce diff between inspect and verify
This commit is purely cosmetics. The goal is to make it easier to read
and compare those two examples (with a diff tool such as meld).

Note that both these examples do exactly the same thing in terms of
reading the MBR and GPT. One of them try and catches exception, the
other does not. Let it it be this way.

Signed-off-by: Arnaud Rebillout <arnaud.rebillout@collabora.com>